% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordtoolioR.analyzeCompetitors.R
\name{keywordtoolioR.analyzeCompetitors.tidy}
\alias{keywordtoolioR.analyzeCompetitors.tidy}
\title{Get Keywordtool.io analyze-competitors request as data tidy data frame}
\usage{
keywordtoolioR.analyzeCompetitors.tidy(
  competitorsWebsite,
  country,
  location,
  language,
  currency,
  network,
  complete = "false"
)
}
\arguments{
\item{competitorsWebsite}{Url of the website you want to get keywords for.}

\item{country}{A country code identifying the country. Complete list can be found here: https://docs.keywordtool.io/reference#keyword-suggestions-google.}

\item{location}{A location code identyfiying the location. Complete list can be found here: https://docs.keywordtool.io/reference#analyze-competitors-google-metrics-location.}

\item{language}{A language code identifying the results language. Complete list can be found here: https://docs.keywordtool.io/reference#analyze-competitors-google-metrics-language.}

\item{currency}{A currency code identifying the results currency. Complete list can be found here: https://docs.keywordtool.io/reference#analyze-competitors-google-metrics-currency.}

\item{network}{"googlesearch" or "googlesearchnetwork" for only search or search and partners respectively.}

\item{complete}{true or false if results should be complete.}
}
\description{
The function gets tidy data frame output for single analyze competitors request.
}
