% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordtoolioR.getQuota.r
\name{keywordtoolioR.getQuota}
\alias{keywordtoolioR.getQuota}
\title{Get Keywordtool.io remaining quota}
\usage{
keywordtoolioR.getQuota()
}
\description{
The function fetches remained keywordtool.io quotas as data frame
}
