% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordtoolioR.getSuggestions.r
\name{keywordtoolioR.getSuggestions.tidy}
\alias{keywordtoolioR.getSuggestions.tidy}
\title{Get Kewordtool.io suggestions single request as tidy data frame}
\usage{
keywordtoolioR.getSuggestions.tidy(
  source,
  keyword,
  country,
  language,
  currency
)
}
\arguments{
\item{source}{One of services ids from the list: google, youtube, bing, amazon, ebay, instagram}

\item{keyword}{A seed keyword you want to check related keywords to.}

\item{country}{A country code identifying the country. Different depending on the source. Complete list can be found here: https://docs.keywordtool.io/reference#keyword-suggestions-google.}

\item{language}{A language code identyfiying the results language. Different depending on the source. Complete list can be found here: https://docs.keywordtool.io/reference#keyword-suggestions-google.}

\item{currency}{A currency code identyfiying the results currency. Different depending on the source. Complete list can be found here: https://docs.keywordtool.io/reference#keyword-suggestions-google.}
}
\description{
The function gets raw data.frame output for single suggestions request.
}
