% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordtoolioR.utils.storeCredentials.r
\name{keywordtoolioR.utils.storeCredentials}
\alias{keywordtoolioR.utils.storeCredentials}
\title{Safely store your keywordtoolio credentials to your global environement.}
\usage{
keywordtoolioR.utils.storeCredentials()
}
\description{
Safely store your keywordtoolio credentials to your global environement.
}
